Set-MailboxDelegation integration testing is complete and the tool is ready for production use
Cleanup some of the comments and formatting
Update the pester test for pipeline input - now all tests will pass

Add -whatif:$false and -confirm:$false to the Remove-Module function
This is necessary because if -session and -whatif are called together, the tool will load the temp module but not unload it since Remove-Module will honor the -whatif flag.  This leaves the temp module loaded in the current session.  Note that Remove-Module invokes another call behind the scenes which still follows the whatif and confirm behavior - gives the message: Performing the operation "Unsubscribe" on target "Event subscription 'Implicit remoting event.  This second call doesn’t honor the flags passed to Remove-Module.  Setting the global $WhatIfPreference = $false and $ConfirmPreference = "none" also fails to suppress this message.  As a result, when using whatif and Confirm, the user will get the above message.

Improve help documentation to clarify the pipeline input must provide a SAM account name or UPN
It turns out that PowerShell coerces the ADUser object (as generated by Get-aduser jsmith, for example) into a distinguished name string - the hybrid tenant is not able to use this input as it cannot do a search against the on-prem domain.  Pipeline input will work OK when provided a SAM account name or UPN however.