Add error handling for failed PSSession connections to the exchange environments.
Prompt the user if they wish to continue with ShouldContinue - this function will return T/F depending on the user input.  When breaking from the function, break seems to be the correct choice.  Exit will close the entire PowerShell console and return only skips the remainder of the BEGIN block.