Implement code to hide the mailbox and code to set mailbox delegation
Preference variables are not passed to module functions like Set-MailboxHidden and Set-MailboxDelegation by default, so need to call them with "-WhatIf:$WhatIfPreference -Verbose:$VerbosePreference" so at least these two will work.  Using parameter splatting so the code is more readable.

Add -WhatIf:$false to the Remove-PSSession calls.  The PSSession establishment is not subject to ShouldProcess since it doesn’t cause a system change, without the  -WhatIf:$false an orphan PSSession will be left in the users session when -whatif is used.