Complete the process block.  The main coding is complete, integration testing needs to be performed.
This new code has not been tested yet:
Implement code to modify the account description with the date, disable the account and move it to the disabled user OU.
Add a simple lookup to display the state of the user account at the end of processing.
Change $user to $userAccount for the Format-UserGroups call to be consistent with the other calls.
Change $user to $userUPN for the calls to the exchange environment.