Finish integration testing and fine tuning Disable-UserAccount
Set the function ConfirmImpact to Medium - when set to High, the remove user groups part of the script always prompts for confirmation since the default ConfirmImpact setting for a PowerShell shell is High.
Remove the verification output and replace it with a simple 'completed' message.  Any errors should be caught by the error handling which will display a warning message something failed.
Update parameter name session to pssession to match upstream changes in Set-MailboxHidden and Set-MailboxDelegation.