Finish the functionality and test the tool - works as expected
Rename the tool to Remove-CryptoBlock to better reflect its purpose.

Update help to note that the SmbShare module requires API's that require Server 2012 or newer.

Remove the ComputerName foreach loop.  The way the foreach user and the invoke-command are nested is a little messy, since Invoke-Command will need to recreate the PSSession for each user in the loop.  However, this implementation was selected because invoking against multiple users won't be common and makes the code simpler by leveraging the capabilities of Invoke-Command.