Refactor lockout behavior
Instead of immediately triggering a lockout on the user, take a graduated approach.  If 3 events happen in 1 minute, lock the user out.  The account will be automatically unlocked after a period of time.  If the user immediately gets locked out again, the script will not automatically schedule an unlock.

Generally restructure the algorithm and add extra logic to handle the new functionality (algorithm)
Copy over some code to handle the email functionality, needs to be adapted for this function