Finish CryptoBlocker and test
Small formatting tweaks to the messages
Need to add .value to the EventProperty objects to get the value of the string
Mod the install directions - set the script to be a triggered scheduled task instead of running from FSRM.  Testing shows fsrm will only run the script the first time it is triggered if there are repeat instances.  However, fsrm will continue logging the event to the windows event log which can be used.
Add task xml file to setup the triggered scheduled task
Update the repo README